+++ soc/arm/microchip_mec/mec1501/soc.h
@@ -14,11 +14,15 @@
 #include "MEC1501hsz.h"
 #include "regaccess.h"
 
+/* common SoC API */
 #include "../common/soc_dt.h"
 #include "../common/soc_gpio.h"
 #include "../common/soc_pins.h"
 #include "../common/soc_espi_channels.h"
 #include "../common/soc_espi_saf.h"
+
+/* common peripheral register defines */
+#include "../common/reg/mec_gpio.h"
 
 #endif
